requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
jinja2>=3.1.0
python-dotenv>=1.0.0
pydantic>=2.9.0
openai>=1.50.0
backoff>=2.2.0
beautifulsoup4>=4.12.0
requests>=2.31.0
numpy>=1.26.0
scipy>=1.12.0
